+++ react/__mocks__/vtex.format-currency.tsx
@@ -1,0 +1,5 @@
+import React, { FunctionComponent } from 'react'
+
+export const FormattedCurrency: FunctionComponent = ({ value }: any) => (
+  <div>{value}</div>
+)