+++ react/__mocks__/vtex.styleguide.tsx
@@ -2,5 +2,6 @@ import Dropdown from '@vtex/styleguide/lib/Dropdown'
 import Input from '@vtex/styleguide/lib/Input'
 import Link from '@vtex/styleguide/lib/Link'
 import Spinner from '@vtex/styleguide/lib/Spinner'
+import IconDelete from '@vtex/styleguide/lib/icon/Delete'
 
-module.exports = { Input, Dropdown, Link, Spinner }
+module.exports = { Input, Dropdown, Link, Spinner, IconDelete }